+++ Hud/Menu/MultipleChildren/MultipleChildWidget.cs
@@ -1,7 +1,7 @@
 ﻿using OpenTK.Mathematics;
 using OpenTK.Windowing.GraphicsLibraryFramework;
 
-namespace Hud.Menu;
+namespace Hud.Menu.MultipleChildren;
 
 public abstract class MultipleChildWidget : Widget
 {
